Update: asignar_actividades_django.py
