+++ gestion_voluntarios/tests/features/asignar_actividades_django.feature
@@ -5,10 +5,13 @@
     quiero que se asignen voluntarios a las actividades,
     para atender una emergencia.
 
-    Esquema del escenario: : Asignar un voluntario a una actividad en una emergencia
+    Esquema del escenario: Asignar un voluntario a una actividad en una emergencia
       #Dado: precondiciones
       Dado que se tiene una emergencia que aún no es atendida
-      Y hay un voluntario que asistirá sin actividades asignadas
+      Y hay una "<cantidad_voluntarios>" que asistirán sin actividades asignadas
+      #  Ejemplos:
+      # | cantidad_voluntarios|
+      # | 2                   |
       Y existe una actividad con un "<nombre_actividad>" sin voluntarios asignados
       #Cuando: eventos
       Cuando se le asigna un voluntario a una actividad existente
@@ -18,5 +21,6 @@
       Y el estado de la emergencia sera"<estado_emergencia>"
 
       Ejemplos: Asignaciones
-      | nombre_actividad    | es_asignado | tiene_voluntario | estado_emergencia |
-      |"Suturar"            | "True"      | "True"           | "True"            |
+      | cantidad_voluntarios | nombre_actividad    | es_asignado | tiene_voluntario | estado_emergencia |
+      | 5                    |"Suturar"            | "True"      | "True"           | "True"            |
+      | 3                    |"Evacuar"            | "True"      | "True"           | "True"            |
